wip: export time range and dataset panels to their own components, so it is easier to edit and maintain. prep for using a caret expanding panel for both
